refactor: return RESTclient form render, move main files into src/components/main folder (#47)
* refactor: move main files to main folder

* fix: fix path trouble for signIn and signUp

* feat: add pre-push linter and typechek

* refactor: change ci pipeline

* refactor: remove test coverage form ci pipeline

* refactor: remove test  from ci pipeline

* refactor: remove ci pipeline

* feat: rename files

* fix: fix path issues

* feat: rename Header

* refactor: rename signUp.test.tsx to SignUp.test.tsx

* refacotr: rename signIn.test.tsx to SignIn.test.tsx

* refactor: update and rename header.test.tsx to Header.test.tsx
