--- .github/workflows/build.yml
@@ -1,34 +1,0 @@
-name: Build
-on:
-  push:
-    branches:
-      - sprint-1
-  pull_request:
-    types: [opened, synchronize, reopened]
-
-jobs:
-  sonarqube:
-    name: SonarQube
-    runs-on: windows-latest
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          fetch-depth: 0
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: 20
-          cache: npm
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Run tests with coverage
-        run: npm run test -- --coverage
-
-      - name: SonarQube Scan
-        uses: SonarSource/sonarqube-scan-action@v5
-        env:
-          SONAR_TOKEN: ${{ secrets.SONAR_TOKEN }}
