+++ .husky/pre-push
@@ -1,1 +1,1 @@
-npm test
+npm run lint && npm run typecheck && npm test
