--- sonar-project.properties
@@ -1,19 +1,0 @@
-sonar.projectKey=Mxmmsv_rest-client-app
-sonar.organization=mxmmsv
-
-sonar.sources=src
-sonar.tests=src
-sonar.test.inclusions=**/*.test.ts,**/*.test.tsx,**/*.spec.ts,**/*.spec.tsx
-sonar.typescript.lcov.reportPaths=coverage/lcov.info
-
-
-# This is the name and version displayed in the SonarCloud UI.
-#sonar.projectName=rest-client-app
-#sonar.projectVersion=1.0
-
-
-# Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-#sonar.sources=.
-
-# Encoding of the source code. Default is default system encoding
-#sonar.sourceEncoding=UTF-8
