+++ src/app/[locale]/(main)/login/page.tsx
@@ -1,4 +1,4 @@
-import SignIn from '@/components/signIn';
+import SignIn from '@/components/main/SignIn';
 
 export default function SignInPage() {
   return <SignIn />;
