+++ src/app/[locale]/(main)/register/page.tsx
@@ -1,4 +1,4 @@
-import SignUp from '@/components/signUp';
+import SignUp from '@/components/main/SignUp';
 
 export default function SignUpPage() {
   return <SignUp />;
