+++ src/app/[locale]/layout.tsx
@@ -7,7 +7,7 @@ import { hasLocale, NextIntlClientProvider } from 'next-intl';
 import 'dayjs/locale/ru';
 
 import Footer from '@/components/footer/Footer';
-import Header from '@/components/header/header';
+import Header from '@/components/header/Header';
 import { routing } from '@/i18n/routing';
 
 const antLocales = { en: enUS, ru: ruRU } as const;
