+++ src/components/main/__tests__/SignIn.test.tsx
@@ -5,8 +5,8 @@ import { vi } from 'vitest';
 
 import { useAuth } from '@/lib/auth/useAuth';
 
-import { mockUser } from '../__mock__/firebaseUser.mock';
-import SignIn from '../signIn';
+import { mockUser } from '../../__mock__/firebaseUser.mock';
+import SignIn from '../SignIn';
 
 vi.mock('react-firebase-hooks/auth', () => ({
   useAuthState: vi.fn(),
