+++ src/components/main/__tests__/SignUp.test.tsx
@@ -5,8 +5,8 @@ import { vi } from 'vitest';
 
 import { useAuth } from '@/lib/auth/useAuth';
 
-import { mockUser } from '../__mock__/firebaseUser.mock';
-import SignUp from '../signUp';
+import { mockUser } from '../../__mock__/firebaseUser.mock';
+import SignUp from '../SignUp';
 
 vi.mock('react-firebase-hooks/auth', () => ({
   useAuthState: vi.fn(),
